backend/requirements.txt
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pypdfium2==4.30.0
python-docx==1.2.0
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
openai==1.54.0
python-dotenv==1.0.0
